Remove methods that are not used
