+++ tests/unit_tests/pose_estimation_metrics_test.py
@@ -12,6 +12,7 @@ import torch.cuda
 from pycocotools.coco import COCO
 from pycocotools.cocoeval import COCOeval
 
+from super_gradients.common.abstractions.abstract_logger import get_logger
 from super_gradients.module_interfaces import PoseEstimationPredictions
 from super_gradients.training.datasets.pose_estimation_datasets.coco_utils import (
     remove_duplicate_annotations,
@@ -19,6 +20,8 @@ from super_gradients.training.datasets.pose_estimation_datasets.coco_utils impor
     remove_crowd_annotations,
 )
 from super_gradients.training.metrics.pose_estimation_metrics import PoseEstimationMetrics
+
+logger = get_logger(__name__)
 
 
 class TestPoseEstimationMetrics(unittest.TestCase):
@@ -205,7 +208,6 @@ class TestPoseEstimationMetrics(unittest.TestCase):
     def _coco_convert_predictions_to_dict(self, predicted_poses, predicted_scores, image_ids):
         kpts = collections.defaultdict(list)
         for poses, scores, image_id_int in zip(predicted_poses, predicted_scores, image_ids):
-
             for person_index, kpt in enumerate(poses):
                 area = (np.max(kpt[:, 0]) - np.min(kpt[:, 0])) * (np.max(kpt[:, 1]) - np.min(kpt[:, 1]))
                 kpt = self._coco_process_keypoints(kpt)
@@ -282,6 +284,90 @@ class TestPoseEstimationMetrics(unittest.TestCase):
                 tmp[i][0:3] = [float(keypoints[i][0]), float(keypoints[i][1]), float(keypoints[i][2])]
 
         return tmp
+
+    @staticmethod
+    def make_keypoints_outside_image_invisible(coco: COCO) -> COCO:
+        for ann in coco.anns.values():
+            keypoints = np.array(ann["keypoints"]).reshape(-1, 3)
+            image_rows = coco.imgs[ann["image_id"]]["height"]
+            image_cols = coco.imgs[ann["image_id"]]["width"]
+
+            visible_joints = keypoints[:, 2] > 0
+            joints_outside_image = (keypoints[:, 0] < 0) | (keypoints[:, 0] >= image_cols) | (keypoints[:, 1] < 0) | (keypoints[:, 1] >= image_rows)
+            visible_joints_outside_image = visible_joints & joints_outside_image
+            if visible_joints_outside_image.any():
+                logger.debug(
+                    f"Detected GT joints outside image size (width={image_cols}, height={image_rows}). "
+                    f"{keypoints[joints_outside_image]} for obj_id {ann['id']} image_id {ann['image_id']}. "
+                    f"Changing visibility to invisible."
+                )
+                keypoints[visible_joints_outside_image, 2] = 0
+
+                ann["keypoints"] = [float(x) for x in keypoints.reshape(-1)]
+        return coco
+
+    @staticmethod
+    def remove_duplicate_annotations(coco: COCO) -> COCO:
+        ann_to_remove = []
+
+        image_ids = list(coco.imgs.keys())
+        for image_id in image_ids:
+            ann_ids = coco.getAnnIds(imgIds=image_id)
+            annotations = coco.loadAnns(ann_ids)
+
+            joints = []
+            for ann in annotations:
+                keypoints = np.array(ann["keypoints"]).reshape(-1, 3)
+                joints.append(keypoints[:, :2])
+
+            if len(joints) == 0:
+                continue
+
+            joints = np.stack(joints, axis=0)
+            gt_joints1 = np.expand_dims(joints, axis=0)  # [1, Num_people, Num_joints, 2]
+            gt_joints2 = np.expand_dims(joints, axis=1)  # [Num_people, 1, Num_joints, 2]
+            diff = np.sqrt(np.sum((gt_joints1 - gt_joints2) ** 2, axis=-1))  # [Num_people, Num_people, Num_joints]
+            diffmean = np.mean(diff, axis=-1)
+
+            duplicate_mask = np.triu(diffmean < 2, k=1)
+            duplicate_indexes_i, duplicate_indexes_j = np.nonzero(duplicate_mask)
+
+            for j in duplicate_indexes_j:
+                ann_to_remove.append(ann_ids[j])
+
+        if len(ann_to_remove) > 0:
+            ann_to_remove = set(ann_to_remove)
+            logger.debug(f"Removing {len(ann_to_remove)} duplicate annotations")
+            len_before = len(coco.dataset["annotations"])
+            coco.dataset["annotations"] = [v for v in coco.dataset["annotations"] if v["id"] not in ann_to_remove]
+            len_after = len(coco.dataset["annotations"])
+            logger.debug(f"Removed {len_before - len_after} duplicate annotations")
+            coco.createIndex()
+
+        return coco
+
+    @staticmethod
+    def remove_crowd_annotations(coco: COCO):
+        ann_to_remove = []
+
+        image_ids = list(coco.imgs.keys())
+        for image_id in image_ids:
+            ann_ids = coco.getAnnIds(imgIds=image_id)
+            annotations = coco.loadAnns(ann_ids)
+
+            for ann in annotations:
+                if bool(ann["iscrowd"]):
+                    ann_to_remove.append(ann["id"])
+
+        if len(ann_to_remove) > 0:
+            logger.debug(f"Removing {len(ann_to_remove)} crowd annotations")
+            len_before = len(coco.dataset["annotations"])
+            coco.dataset["annotations"] = [v for v in coco.dataset["annotations"] if v["id"] not in ann_to_remove]
+            len_after = len(coco.dataset["annotations"])
+            logger.debug(f"Removed {len_before - len_after} crowd annotations")
+            coco.createIndex()
+
+        return coco
 
 
 if __name__ == "__main__":
